tests/wpt/web-platform-tests/xhr/firing-events-http-no-content-length.html

<!doctype html>
<html>
  <head>
    <title>ProgressEvent: firing events for HTTP with no Content-Length</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#firing-events-using-the-progressevent-interface">
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test();

      test.step(function() {
        var xhr = new XMLHttpRequest();

        xhr.onprogress = function(pe) {
          test.step(function() {
            if(pe.type == "progress") {
              assert_greater_than_equal(pe.loaded, 0, "loaded");
              assert_false(pe.lengthComputable, "lengthComputable");
              assert_equals(pe.total, 0, "total");
            }
          }, "Check lengthComputed, loaded, total when Content-Length is NOT given.");
        }

        // "loadstart", "error", "abort", "load" tests are out of scope.
        // They SHOULD be tested in each spec that implement ProgressEvent.

        xhr.onloadend = function(pe) {
          test.done();
        }
        xhr.open("GET", "resources/trickle.py?ms=0&count=100", true);
        xhr.send(null);
      })
    </script>
  </body>
</html>
